#165 выбор цвета для графика PIE
* add color picker for bar chart elements
