+++ packages/charts-editor/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@seine/charts-editor",
-  "version": "0.3.20",
+  "version": "0.3.21",
   "description": "Seine charts editor.",
   "main": "lib/cjs/index.js",
   "module": "lib/esm/index.js",
