+++ packages/content/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@seine/content",
-  "version": "0.3.21",
+  "version": "0.3.22",
   "description": "Seine content renderer.",
   "main": "lib/cjs/index.js",
   "module": "lib/esm/index.js",
