+++ packages/ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@seine/ui",
-  "version": "0.3.20",
+  "version": "0.3.21",
   "description": "Seine UI library.",
   "main": "lib/cjs/index.js",
   "module": "lib/esm/index.js",
